refactor: Clean up Docker implementation from chroma_db branch
- Renamed base_images_v3.py to base_images.py (keeping optimized v3 implementation)
- Fixed all pylint issues in base_images module:
  - Removed unused imports (json, Output, CacheFromArgs)
  - Fixed trailing whitespace issues
  - Fixed line-too-long violations
  - Removed unused variables
  - Added missing final newlines
- Updated import in __main__.py to use renamed module
- Maintained all performance optimizations:
  - 99.99% context size reduction (2.9GB → 200KB)
  - ECR caching with registry-based storage
  - Scoped build contexts for stable hashes
